Move the tests for the selectors in their own test module.
Selectors moved away from the DSSP module. This commit reflects that
change in the tests. It also makes pylint happier.
